+++ detekt-rules/src/main/kotlin/io/gitlab/arturbosch/detekt/rules/providers/CommentSmellProvider.kt
@@ -1,8 +1,12 @@
-package io.gitlab.arturbosch.detekt.rules.documentation
+package io.gitlab.arturbosch.detekt.rules.providers
 
 import io.gitlab.arturbosch.detekt.api.Config
 import io.gitlab.arturbosch.detekt.api.RuleSet
 import io.gitlab.arturbosch.detekt.api.RuleSetProvider
+import io.gitlab.arturbosch.detekt.rules.documentation.CommentOverPrivateMethod
+import io.gitlab.arturbosch.detekt.rules.documentation.CommentOverPrivateProperty
+import io.gitlab.arturbosch.detekt.rules.documentation.UndocumentedPublicClass
+import io.gitlab.arturbosch.detekt.rules.documentation.UndocumentedPublicFunction
 
 /**
  * @author Artur Bosch
