+++ detekt-rules/src/main/kotlin/io/gitlab/arturbosch/detekt/rules/providers/ComplexityProvider.kt
@@ -1,8 +1,15 @@
-package io.gitlab.arturbosch.detekt.rules.complexity
+package io.gitlab.arturbosch.detekt.rules.providers
 
 import io.gitlab.arturbosch.detekt.api.Config
 import io.gitlab.arturbosch.detekt.api.RuleSet
 import io.gitlab.arturbosch.detekt.api.RuleSetProvider
+import io.gitlab.arturbosch.detekt.rules.complexity.ComplexMethod
+import io.gitlab.arturbosch.detekt.rules.complexity.LargeClass
+import io.gitlab.arturbosch.detekt.rules.complexity.LongMethod
+import io.gitlab.arturbosch.detekt.rules.complexity.LongParameterList
+import io.gitlab.arturbosch.detekt.rules.complexity.NestedBlockDepth
+import io.gitlab.arturbosch.detekt.rules.complexity.TooManyFunctions
+import io.gitlab.arturbosch.detekt.rules.complexity.ComplexCondition
 
 /**
  * @author Artur Bosch
