+++ detekt-rules/src/main/kotlin/io/gitlab/arturbosch/detekt/rules/providers/StyleGuideProvider.kt
@@ -1,8 +1,11 @@
-package io.gitlab.arturbosch.detekt.rules.style
+package io.gitlab.arturbosch.detekt.rules.providers
 
 import io.gitlab.arturbosch.detekt.api.Config
 import io.gitlab.arturbosch.detekt.api.RuleSet
 import io.gitlab.arturbosch.detekt.api.RuleSetProvider
+import io.gitlab.arturbosch.detekt.rules.style.MaxLineLength
+import io.gitlab.arturbosch.detekt.rules.style.NamingConventionViolation
+import io.gitlab.arturbosch.detekt.rules.style.WildcardImport
 
 /**
  * @author Artur Bosch
