+++ detekt-rules/src/main/resources/META-INF/services/io.gitlab.arturbosch.detekt.api.RuleSetProvider
@@ -1,7 +1,7 @@
-io.gitlab.arturbosch.detekt.rules.complexity.ComplexityProvider
+io.gitlab.arturbosch.detekt.rules.providers.ComplexityProvider
 io.gitlab.arturbosch.detekt.rules.providers.CodeSmellProvider
-io.gitlab.arturbosch.detekt.rules.style.StyleGuideProvider
-io.gitlab.arturbosch.detekt.rules.documentation.CommentSmellProvider
+io.gitlab.arturbosch.detekt.rules.providers.StyleGuideProvider
+io.gitlab.arturbosch.detekt.rules.providers.CommentSmellProvider
 io.gitlab.arturbosch.detekt.rules.providers.EmptyCodeProvider
 io.gitlab.arturbosch.detekt.rules.providers.PotentialBugProvider
 io.gitlab.arturbosch.detekt.rules.providers.ExceptionsProvider
